backend/requirements.txt
Django>=5.0,<6.0
 djangorestframework>=3.15
 djangorestframework-authtoken>=1.2
 django-cors-headers>=4.0
 PyMySQL>=1.0.2
